fix: Adds /fg/assets/ to nginx.conf
This adds a /fg/assets/ location to the nginx server, so that we can
serve the fileglancer static assets from a reliable location. Before
this change, the assets were being referenced via a relative url, eg
"./assets/index.js". This worked as long as we didn't have a url of more
than one level deep. For example, urls like /fg/browse and /fg/help worked,
but if we needed to have a url like /fg/browse/file/data, the assets
request url would be /fg/browse/file/assets, instead of /fg/assets.
Typically this could be fixed by setting a base in the vite config file,
but we have the added complexity that we are using jupyterhub as a
reverse proxy, with the user name in the url, so we need to serve the
site from /user/<username>/fg. This would be different everytime, so the
base could not be added to the index.html file at build time. Since we
are using nginx in front of the jupyterhub server, we can grab the
requests for /fg/assets and serve those requests directly from the
filesystem. This also saves us from using jupyterhub to serve up static
assets.